backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dnspython==2.8.0
fastapi==0.110.1
h11==0.16.0
idna==3.11
motor==3.3.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.5.0
python-dotenv==1.2.1
requests==2.32.5
starlette==0.37.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.25.0
